requirements.txt
absl-py
gym
numpy
pybullet
matplotlib
opencv-python
meshcat
scipy
scikit-image
torch
hydra-core
ftfy
torchvision
transforms3d
wandb
kornia
gym
pytorch-lightning
transformers
meshcat
openai
seaborn
trimesh
rtree